Optimized the next and previous functions.
Changed the funcions that get the next row for a given row. Before,
they would construct a list of all of the non-directory items and
then go one down or one up. I wrote it like that because I didn't
want to think about how to do it well but it meant when there were
more than a few hundred files that it would hang becuaes it had to
construct a list every time. Now it correctly finds the next one
much faster.