Changed the minimum required gtk version to 3.18
I changed it because I tried compiling on Ubuntu but it didn't work
becuase it seems like glade set the minimum version to 3.20. I'm
not sure if that's a good thing but that means it wouldn't run on
Ubuntu. Manually changed the number to 3.18 in the files but I
don't even know if that's low enough is some systems have older
versions.